Rename providers/upgrade/ to providers/recompute/ and update all imports
