+++ providers/index.ts
@@ -19,9 +19,9 @@ import azblobConfig from '../providers/stores/azblobConfig.ts'
 import dispatchConfig from '../providers/stores/dispatchConfig.ts'
 import fileConfig from '../providers/stores/fileConfig.ts'
 import mongoConfig from '../providers/stores/mongoConfig.ts'
-import upgradeAzureQueueConfig from '../providers/upgrade/azureQueueConfig.ts'
-import upgradeMemoryQueueConfig from '../providers/upgrade/memoryQueueConfig.ts'
-import * as recomputeHandler from '../providers/upgrade/recomputeHandler.ts'
+import upgradeAzureQueueConfig from '../providers/recompute/azureQueueConfig.ts'
+import upgradeMemoryQueueConfig from '../providers/recompute/memoryQueueConfig.ts'
+import * as recomputeHandler from '../providers/recompute/recomputeHandler.ts'
 import type { ICache } from './caching/index.js'
 import listBasedFilterConfig from './harvest/throttling/listBasedFilterConfig.ts'
 import type { Logger } from './logging/index.js'
