+++ test/providers/recompute/onDemandComputePolicy.ts
@@ -10,7 +10,7 @@ import { expect } from 'chai'
 import sinon from 'sinon'
 import type { RecomputeContext } from '../../../business/definitionService.ts'
 import EntityCoordinates from '../../../lib/entityCoordinates.ts'
-import { OnDemandComputePolicy } from '../../../providers/upgrade/onDemandComputePolicy.ts'
+import { OnDemandComputePolicy } from '../../../providers/recompute/onDemandComputePolicy.ts'
 
 describe('OnDemandComputePolicy', () => {
   let policy
