+++ test/providers/recompute/recomputeHandler.ts
@@ -12,12 +12,12 @@ import type { Definition, DefinitionService, RecomputeContext } from '../../../b
 import DefinitionServiceFactory from '../../../business/definitionService.ts'
 import EntityCoordinates from '../../../lib/entityCoordinates.ts'
 import type { IQueue } from '../../../providers/queueing/index.js'
-import DefinitionQueueUpgrader from '../../../providers/upgrade/defUpgradeQueue.ts'
-import { DefinitionVersionChecker } from '../../../providers/upgrade/defVersionCheck.ts'
-import { DelayedComputePolicy } from '../../../providers/upgrade/delayedComputePolicy.ts'
-import memoryQueueConfig from '../../../providers/upgrade/memoryQueueConfig.ts'
-import { OnDemandComputePolicy } from '../../../providers/upgrade/onDemandComputePolicy.ts'
-import { defaultFactory, delayedFactory, RecomputeHandler } from '../../../providers/upgrade/recomputeHandler.ts'
+import DefinitionQueueUpgrader from '../../../providers/recompute/defUpgradeQueue.ts'
+import { DefinitionVersionChecker } from '../../../providers/recompute/defVersionCheck.ts'
+import { DelayedComputePolicy } from '../../../providers/recompute/delayedComputePolicy.ts'
+import memoryQueueConfig from '../../../providers/recompute/memoryQueueConfig.ts'
+import { OnDemandComputePolicy } from '../../../providers/recompute/onDemandComputePolicy.ts'
+import { defaultFactory, delayedFactory, RecomputeHandler } from '../../../providers/recompute/recomputeHandler.ts'
 import { createMockLogger, createSilentLogger, type StubbedLogger } from '../../helpers/mockLogger.ts'
 
 const TEST_COORDINATES = 'npm/npmjs/-/leftpad/1.0.0'
